// src/app/taf-fiscal/models/item-table-details-resources-passed-on-the-by-the-sports-association.ts
export interface ItemTableDetailsResourcesPassedOnTheByTheSportsAssociation {
  branchId: string;
  taxNumber?: string;
  taxNumberFormated?: string;
  company?: string;
  totalTransferAmount?: number;
  totalGrossValue?: number;
  totalTaxBase?: number;
  totalTaxes?: number;
  totalReceived?: number;
  totalValueOfRetentionWithSuspendedLiability?: number;
  key?: string;
  grossValue?: number;
  item?: string;
  invoiceKey?: string;
  sourceTaxNumber: string;
  associationTaxNumber?: string;
}